backend/requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
Pillow
opencv-python
pyyaml
pdfminer.six
rapidocr-onnxruntime
python-dateutil
psutil
pydantic
requests
pyjwt
cryptography
